+++ src/app/contact/page.tsx
@@ -1,11 +1,13 @@
-import styles from "@/styles/common.module.css";
+import styles from "@/styles/contact/contactpage.module.css";
+import { homeConstants } from "@/constants/constants";
+import ContactContent from "@/container/contact/ContactContent";
 
 export default function ContactPage() {
   return (
-    <div id="contact" className={styles.commonContainer}>
-      <div>phone: 010-0000-0000</div>
-      <div>email: 00@naver.com</div>
-      <div>mailing section</div>
+    <div id="contact" className={styles.contactContainer}>
+      <ContactContent />
+      <div>phone: {homeConstants.PHONE_NUM}</div>
+      <div>email: {homeConstants.EMAIL_ADD}</div>
     </div>
   );
 }
