+++ src/styles/common.module.css
@@ -9,7 +9,4 @@
   /* for test */
   background-color: rgba(248, 247, 245, 0.9);
   border-bottom: 1px solid white;
-
-  /* text */
-  color: white;
 }
