+++ src/styles/components/footer.module.css
@@ -1,7 +1,5 @@
 .footer {
   display: flex;
   justify-content: center;
-  margin: 10px;
-
-  color: white;
+  padding: 30px;
 }
