+++ src/styles/contact/contactcontent.module.css
@@ -1,0 +1,50 @@
+.contactContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.inputGroup {
+  position: relative;
+  margin-top: 10px;
+}
+
+.inputGroup input,
+.inputGroup textarea {
+  font-size: 15px;
+  border: none;
+  border-bottom: 1px solid black;
+  padding-bottom: 20px;
+  text-align: center;
+  position: relative;
+  background: none;
+  z-index: 5;
+
+  &:focus {
+    outline: none;
+  }
+
+  &::placeholder {
+    color: black;
+  }
+}
+
+.inputGroup span {
+  display: block;
+  position: absolute;
+  bottom: 0;
+  left: 0%;
+  background-color: #666;
+  width: 0;
+  height: 2px;
+  border-radius: 2px;
+  transition: 0.5s;
+}
+
+.inputGroup label {
+  position: absolute;
+  color: #aaa;
+  font-size: 20px;
+  bottom: 8px;
+  transition: all 0.2s;
+}
