+++ src/styles/contact/contactpage.module.css
@@ -1,0 +1,12 @@
+.contactContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+  width: 100vw;
+}
+
+.homeContent {
+  flex: 1;
+}
