+++ src/styles/page.module.css
@@ -1,0 +1,3 @@
+.main {
+  background-color: rgba(248, 247, 245, 0.9);
+}
